Switch away from coroutine based multimsg
ZeroMQ reserves the memory for all message parts anyway, thus there is
no memory advantage in the previous coroutine based solution. It also
causes other problems, for example validation of recieved messages can
be significantly simplified if the number of message parts is known at
the start, which is not possible when using coroutines, at least not
without consuming the entire coroutine which is basically an indirect
form of this new solution.