Add static methods building objects
We generally need to do some complex operations when constructing
message objects, such as searching the message parts to collect
multiple message parts that all belong to the same section of the
message. We can't do this in the constructor of the message, because
it would be very difficult and inefficient to do these complex
operations in the initializer list itself. Instead, we can do these
operations in a static method, then pass the message sections to a
simple constructor that simply moves the sections into the object.

The downside to this solution is that the classes become rather
verbose with big constructors that accept a lot of parameters.