Finalize message API
Message API has been finalized on classes that read message parts from
an iterator using helper functions such as `read_part` and
`read_optional`. The API is somewhat verbose due to the constructors
that just repeat the parts of the message, but it is efficient and
typesafe.

The API may be extended later with a base message class that contains
common operations, such as sending the message.