Implement broker, fix message & socket classes
Implemented the broker's functions for processing messages. In the
meantime, discovered and fixed a few missing functionalities and bugs
on the message classes and the socket.

Message classes now have functions to access their addresses where
necessery.

The `send` function used to return a tuple of iterators. However, this
meant that the coroutine used to create those iterators would go out
of scope, which resulted in iterators becoming invalid. This has been
fixed by returning the coroutine itself.

Following the above fix, the socket class' `send_multimsg` has been
changed to accept any sort of container that can be iterated on.