Fix `msg::send` possibly reading deleted messages
The function `msg::send` accepts `message && msg` as a
parameter. However, The lambda created in the function only captures a
reference. This means that the message doesn't actually get moved
anywhere, and if the message object gets deleted then `send` will try
to read the deleted message.

This has been fixed by making `send` strict; instead of creating a
coroutine it now creates a vector and puts all the message parts
inside that.

This change also removes our dependency on Boost Context.