Fix broker sending work to dead workers
The broker will now check the last time it has seen the worker to make
sure it has not timed out. Note that this is the only point where we
clean up dead workers, which means that a worker that died and has not
been requested will remain registered in the broker. It is safe to
assume that this will be a rare case.

closes #3