Add special exception for empty messages
If `recv_multimsg` times out and the caller tries to use the result
without checking, we end up trying to read an empty function. A
special exception has been added to handle this, to avoid throwing a
generic "expected part is missing" exception which can be confusing,
especially during testing.