Add datastore workers' declarations
Added declarations for datastore readers and writers. Also added the
implementation for a thread-safe singleton that will hold a reference
to an LMDB environment, which needs to be shared across all workers
within the same process.

A C++ wrapper for LMDB has also been added into the project as a
submodule.