Refactor datastore
Refactored datastore modules. The biggest change is that instead of
per-bucket writers, there is now a single writer. The previous desgin
was the result of a misunderstanding of the way LMDB operated.

The `datastore` namespace has been renamed into `data`, and there is
now a `datastore` class that acts as a base class for other datastore
modules.