Fix lmdb++ segfaults & memory issues
When used with strings, lmdb++ misuses the classes by casting them,
and ends up causing segfaults.

In order to avoid this, we instead copy the data into `lmdb::val`
buffers before passing it to lmdb++. This means we have to copy data
forward and back. Luckily, the only point where this might cause a
significant performance drop is the request data being copied from
`lmdb::val` to a string during the read operation.