Add write & read tests
The datastore test now tests both the writer and the reader. This
required us to make the request structs available to the testing code
so that the dynamically created key can be inserted into the
request (the previous method where we embedded the serialized request
into the code wouldn't work).

However, doing this change caused ambiguity with the name `socket`,
thus we had to clarify some of them as `class socket`.